apps/bot/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-stop-signal"
version = "0.1.0"
description = "KRX relative drop trailing stop Telegram bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "APScheduler>=3.10,<4.0",
  "pykrx>=1.0.50",
  "requests>=2.32.0",
  "pandas>=2.2,<3.0",
  "exchange_calendars>=4.5,<5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
packages = ["sss"]

[tool.pytest.ini_options]
testpaths = ["tests"]
